Add “used by” to docs
