+++ docs/src/content/docs/about.md
@@ -3,7 +3,24 @@ title: About
 description: Additional info about this project
 ---
 
-## Project Goals
+## Used by
+
+- [**Bigcommerce**](https://github.com/bigcommerce/bigcommerce-api-node): Node SDK for the BigCommerce API
+- [**Budibase**](https://github.com/Budibase/budibase): low code platform for creating internal tools, workflows, and admin panels
+- [**Fedora `fmn`**](https://github.com/fedora-infra/fmn): tools and APIs for Fedora’s messaging infra
+- [**Fingerprint**](https://github.com/fingerprintjs/fingerprintjs-pro-server-api-node-sdk): device fingerprinting for high-scale applications
+- [**Google Firebase CLI**](https://github.com/firebase/firebase-tools): Official CLI for Google’s Firebase platform
+- [**GitHub Octokit**](https://github.com/octokit): Official SDK for the GitHub API
+- [**Lotus**](https://github.com/uselotus/lotus): open source pricing & packaging infra
+- [**Jitsu**](https://github.com/jitsucom/jitsu): modern, open source data ingestion / data pipelines
+- [**Medusa**](https://github.com/medusajs/medusa): building blocks for digital commerce
+- [**Nitro engine**](https://github.com/unjs/nitro): universal web server engine that powers Nuxt
+- [**Relevance AI**](https://github.com/RelevanceAI/relevance-js-sdk): build and deploy AI chains
+- [**Revolt**](https://github.com/revoltchat/api): open source user-first chat platform
+- [**Spacebar**](https://github.com/spacebarchat): a free, open source, self-hostable Discord-compatible chat/voice/video platform
+- [**Twitter API**](https://github.com/twitterdev/twitter-api-typescript-sdk): Official SDK for the Twitter API
+
+## Project goals
 
 1. Support converting any valid OpenAPI schema to TypeScript types, no matter how complicated.
 1. This library does **NOT** validate your schema, there are other libraries for that.
